Fix: Prevented aggressive-indent from indenting whilst typing new lines
This is a long term issue but it was implimented primarily to fix
unbalanced pairs from screwing with indentation and distracting from typing.

It simply functions by checking if the cursor is less than 3 characters away
from the end of the line, if it is, then it is likely a new line being written,
or extra characters on an existing line.