Documentation: Updated all variables to use bind-key for doc generation
It is  uncertain where  the so  called "vanilla-keys" should  be bound,  that is
consistent with  the rest of  the config.  For now it  will be bound  within the
'bind-keys'  use-package macro,  although rather  annoyingly 'cemacs-init-setup'
still has  to reference this function  with little knowledge of  use-package, as
well as just outright failing if use-package fails...

Ideally  this should  have  a fallback  binding,  but this  works  for now.  The
function call is protected from errors so it doesn't stall futhur initialization
due to errors.

Fix: Cleaned up few warnings and  errors, primarily to do with use-package usage
As it turns out the ':require:'  use-packag form doesn't appear to do dependency
resolution.
This is  resolved by just  redfining packages  as a use-package  declaration and
loading thep primary package ':after' the dependency.